+++ packages/components/radio/src/RadioGroup.styles.ts
@@ -1,0 +1,12 @@
+import { cva, VariantProps } from 'class-variance-authority'
+
+export const radioGroupStyles = cva(['gap-xl flex'], {
+  variants: {
+    orientation: {
+      horizontal: 'flex-row',
+      vertical: 'flex-col',
+    },
+  },
+})
+
+export type RadioGroupVariantsProps = VariantProps<typeof radioGroupStyles>