+++ packages/components/switch/src/Switch.tsx
@@ -1,11 +1,11 @@
-import React from 'react'
+import { forwardRef } from 'react'
 
 import { Input, type InputProps } from './SwitchInput'
 import { Label } from './SwitchLabel'
 
 export type SwitchProps = InputProps
 
-export const Switch = React.forwardRef<HTMLButtonElement, SwitchProps>(
+export const Switch = forwardRef<HTMLButtonElement, SwitchProps>(
   ({ value = 'on', size = 'md', children, className, ...rest }, ref) => {
     return (
       <Label data-spark-component="switch" disabled={rest.disabled} className={className}>