+++ packages/components/badge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7](https://github.com/adevinta/spark/compare/@spark-ui/badge@0.1.6...@spark-ui/badge@0.1.7) (2023-05-25)
+
+**Note:** Version bump only for package @spark-ui/badge
+
+## [0.1.6](https://github.com/adevinta/spark/compare/@spark-ui/badge@0.1.5...@spark-ui/badge@0.1.6) (2023-05-22)
+
+**Note:** Version bump only for package @spark-ui/badge
+
 ## [0.1.5](https://github.com/adevinta/spark/compare/@spark-ui/badge@0.1.4...@spark-ui/badge@0.1.5) (2023-05-11)
 
 **Note:** Version bump only for package @spark-ui/badge