+++ packages/components/icon-button/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.10](https://github.com/adevinta/spark/compare/@spark-ui/icon-button@1.2.9...@spark-ui/icon-button@1.2.10) (2023-05-25)
+
+**Note:** Version bump only for package @spark-ui/icon-button
+
 ## [1.2.9](https://github.com/adevinta/spark/compare/@spark-ui/icon-button@1.2.8...@spark-ui/icon-button@1.2.9) (2023-05-12)
 
 **Note:** Version bump only for package @spark-ui/icon-button